% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikdown.R
\name{loglik.down}
\alias{loglik.down}
\title{Compute the likelihood for the downstream model.}
\usage{
loglik.down(wt, ko, par)
}
\arguments{
\item{wt}{Observationnal data, matrix of dimension N*2, where N stand for the number of replicate.
The first column correspond to the knocked-out gene.}

\item{ko}{Interventionnal data, matrix of dimension N*2, where N stand for the number of replicate.
The first column correspond to the knocked-out gene.}

\item{par}{Set of parameter, such that the first correspond to alpha,
the second and third to the mean, and the last two ones to the log of the standard deviation}
}
\value{
The value of the log likelihood for the downstream model.
}
\description{
Compute the likelihood for the downstream model.
}
\examples{
wt <- rnorm(n=10,mean=-2,sd=0.8)
wt <- cbind(wt,2*wt+rnorm(n=10,mean=1,sd=1.7))
ko <- rnorm(n=10,mean=0,sd=0.01)
ko <- cbind(wt,2*wt+rnorm(n=10,mean=1,sd=1.7))
par <- c(-2,1,0.3,-4,10)
loglik.down(wt,ko,par)
}

