% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcausal.R
\name{post.causal}
\alias{post.causal}
\title{Compute bayes factor associated with a marginal linear causal model.}
\usage{
post.causal(wt, ko, par = c(0, apply(wt, 2, mean), log(apply(wt, 2, sd))),
  gradient = FALSE, prior = 0.5)
}
\arguments{
\item{wt}{Observationnal data, matrix of dimension N*2, where N stand for the number of replicate.
The first column correspond to the knocked-out gene.}

\item{ko}{Interventionnal data, matrix of dimension N*2, where N stand for the number of replicate.
The first column correspond to the knocked-out gene.}

\item{par}{Set of initial parameter for the optimisation, such that the first correspond to alpha,
the second and third to the mean, and the last two ones to the log of the standard deviation}

\item{gradient}{Do we use the gradient? Default set to TRUE.}

\item{prior}{Do we want to promote one model? Default to 0.5.}
}
\value{
The return of the two numerical optimisation, with the optimal value for each parameters, and the value of the bayes factor.
}
\description{
Compute bayes factor associated with a marginal linear causal model.
}
\examples{
wt <- rnorm(n=10,mean=-2,sd=0.8)
wt <- cbind(wt,2*wt+rnorm(n=10,mean=1,sd=1.7))
ko <- rnorm(n=10,mean=0,sd=0.01)
ko <- cbind(wt,2*wt+rnorm(n=10,mean=1,sd=1.7))
par <- c(-2,1,0.3,-4,10)
post.causal(wt,ko,par)$B
}

